moved MockHadoopTestCase to tests.mockhadoop, partially ported test_bin to hadoop runner
